Improve documentation

+++ mapc_optimal/utils.py
@@ -1,8 +1,11 @@
-"""Utility functions, including the function for calculation of the path loss from node positions using
-    the TGax channel model."""
+"""
+Utility functions, including the function for calculation of the path loss from node positions using
+the TGax channel model.
+"""
 
 import numpy as np
 from numpy.typing import NDArray, ArrayLike
+
 from mapc_optimal.constants import BREAKING_POINT, CENTRAL_FREQUENCY, REFERENCE_DISTANCE, WALL_LOSS
 
 
@@ -21,7 +24,7 @@ def dbm_to_lin(x: ArrayLike) -> NDArray:
         Output in a linear scale.
     """
 
-    return np.power(10., x / 10.).astype(float)
+    return np.power(10., x / 10.)
 
 
 def lin_to_dbm(x: ArrayLike) -> NDArray:
@@ -39,7 +42,7 @@ def lin_to_dbm(x: ArrayLike) -> NDArray:
         Output in dBm.
     """
 
-    return 10. * np.log10(x).astype(float)
+    return 10. * np.log10(x)
 
 
 def tgax_path_loss(distance: ArrayLike, walls: ArrayLike) -> NDArray:
